Allow reading swipe data without using stdin (#7)
Adds new configuration settings:

```
[reader]
mode = 'stdin' | 'rawkbd'
device = path
```

With `mode` set to `stdin`, the program behaves as it did before. When mode is set to `rawkbd`, and device is set to a correct keyboard device file (found in `/dev/input/`), the program will not need keyboard input over stdin. This will allow us to set up Gatekeeper as a systemd service instead of auto-starting on login (meaning we get auto-restart on crash!)

Implementation overview:
- Remove `swipe.py`, `Account` now directly receives `id` and `lcc`
- Add `cardreader.py`:
  - Contains `StdinReader` and `RawKbdReader`
  - `StdinReader` functions as before
  - `RawKbdReader` is the new behavior
    - This one is a *little* cursed, parsing linux keyboard events is weird...but it works
  - Both have an `events()` method that returns an iterator of `SwipeEvent`s, which contain ID and LCC information.
- Refactored `main.py` to use the new cardreader code. Calls `get_cardreader` to get the proper reader class, and then iterates over the events.